refactor: 重组后端目录结构
- shared/ → domain/（models.py + schemas.py 合并为单文件）
- crawler/core/ 扁平化（discovery.py, processor.py, database.py, rate_limit.py）
- crawler/api/models/ 合并为单文件 models.py
- app/crud/ 合并为单文件 crud.py
- app/api/v1/ 扁平化为 app/api/
- 删除 core/config.py，crawler/config.py 自行 load_dotenv
- 删除 app/database.py，直接从 domain 导入
