+++ backend/src/app/api/videos.py
@@ -3,10 +3,10 @@ import logging
 from fastapi import APIRouter, Depends, HTTPException
 from sqlmodel import Session
 
-from shared.database import get_session
-from shared.schemas import VideoSchema
+from domain.database import get_session
+from domain.schemas import VideoSchema
 
-from ...crud import crud_video
+from ..crud import get_videos
 
 
 logger = logging.getLogger(__name__)
@@ -17,7 +17,7 @@ router = APIRouter()
 def read_videos(session: Session = Depends(get_session)):
     """获取所有视频列表"""
     try:
-        return crud_video.get_videos(session, is_touhou=False)
+        return get_videos(session, is_touhou=False)
     except Exception as e:
         logger.error(f"Error fetching videos: {e}")
         raise HTTPException(status_code=500, detail="Error fetching videos")
@@ -26,7 +26,7 @@ def read_videos(session: Session = Depends(get_session)):
 def read_touhou_videos(session: Session = Depends(get_session)):
     """只获取东方Project相关的视频"""
     try:
-        return crud_video.get_videos(session, is_touhou=True)
+        return get_videos(session, is_touhou=True)
     except Exception as e:
         logger.error(f"Error fetching touhou videos: {e}")
         raise HTTPException(status_code=500, detail="Error fetching touhou videos")
