--- backend/src/app/database.py
@@ -1,4 +1,0 @@
-from shared.database import get_session
-
-
-__all__ = ["get_session"]
