+++ backend/src/app/main.py
@@ -1,7 +1,7 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
-from .api.v1 import videos, users
+from .api import videos, users
 
 app = FastAPI(
     title="Touhou Memory Archive API",
