--- backend/src/crawler/api/models/__init__.py
@@ -1,24 +1,0 @@
-from .video_detail import VideoDetailData, Page, Owner, VideoStat
-from .video_tag import VideoTag
-from .user_space import (
-    SpaceVideoItem, SpaceList, SpacePageInfo, UserSpaceData,
-    AppSpaceItem, AppSpaceData,
-)
-from .season_archives import SeasonArchiveItem, SeasonMeta, SeasonPageInfo, SeasonArchivesData
-
-__all__ = [
-    "VideoDetailData",
-    "Page",
-    "Owner",
-    "VideoStat",
-    "VideoTag",
-    "SpaceVideoItem",
-    "SpacePageInfo",
-    "UserSpaceData",
-    "AppSpaceItem",
-    "AppSpaceData",
-    "SeasonArchiveItem",
-    "SeasonMeta",
-    "SeasonPageInfo",
-    "SeasonArchivesData",
-]
