+++ backend/src/crawler/config.py
@@ -1,4 +1,10 @@
 import os
+from pathlib import Path
+
+from dotenv import load_dotenv
+
+load_dotenv(Path(__file__).parent.parent.parent / ".env")
+
 # --- 全局行为控制 ---
 # 全局最大并发数：所有网络请求、数据库操作都受此限制(fetch_video_list除外)
 MAX_CONCURRENCY = 5
