+++ backend/src/crawler/main.py
@@ -3,15 +3,15 @@ import logging
 
 import aiohttp
 
-from shared.database import init_db
-from shared.schemas import VideoSchema
+from domain.database import init_db
+from domain.schemas import VideoSchema
 
 from crawler.api.bili_api import BiliAPI
-from crawler.core.video_discovery import VideoDiscovery
+from crawler.discovery import VideoDiscovery
 from .config import MAX_CONCURRENCY, MAX_QUEUE_SIZE
-from crawler.core.database import get_all_user_mids
-from crawler.core.delay_manager import DelayManager
-from crawler.core.video_processor import VideoService
+from crawler.database import get_all_user_mids
+from crawler.rate_limit import DelayManager
+from crawler.processor import VideoService
 
 
 logger = logging.getLogger(__name__)
