+++ backend/src/domain/models.py
@@ -4,7 +4,14 @@ from sqlalchemy.orm import Mapped
 from sqlmodel import SQLModel, Field, Relationship
 
 if TYPE_CHECKING:
-    from shared.schemas.video import VideoSchema
+    from domain.schemas import VideoSchema
+
+
+class User(SQLModel, table=True):
+    __tablename__: str = "users"
+
+    mid: int = Field(primary_key=True)
+    name: str
 
 
 class VideoPart(SQLModel, table=True):
@@ -83,7 +90,7 @@ class Video(SQLModel, table=True):
 
     def to_schema(self, uploader_name: str = "", parts: list | None = None):
         """表模型 → Pydantic schema"""
-        from shared.schemas.video import VideoSchema
+        from domain.schemas import VideoSchema
 
         return VideoSchema(
             aid=self.aid,
