--- backend/src/shared/models/__init__.py
@@ -1,4 +1,0 @@
-from .video import Video, VideoPart
-from .user import User
-
-__all__ = ["Video", "VideoPart", "User"]
