--- backend/src/shared/schemas/__init__.py
@@ -1,4 +1,0 @@
-from .video import VideoSchema, VideoPartSchema
-from .user import UserSchema
-
-__all__ = ["VideoSchema", "VideoPartSchema", "UserSchema"]
