--- backend/src/shared/schemas/user.py
@@ -1,6 +1,0 @@
-from pydantic import BaseModel
-
-
-class UserSchema(BaseModel):
-    mid: int
-    name: str
